refactor: extract bundle verification into application/bundle submodule
First step of splitting the 1704-line application.rs (SRP violation).
Extracts bundle verification, manifest extraction, and path validation
into a dedicated `bundle` submodule with pure functions:

- validate_path_component / validate_artifact_path
- verify_and_extract_manifest (strict, signature required)
- extract_manifest_allow_unsigned (dev mode, signature optional)
- extract_bundle_manifest (parse tar.gz + JSON manifest)
- is_bundle_blob / is_bundle_archive

The original functions remain on NodeClient for backward compatibility.
Future PRs can migrate callers to use the bundle module directly.
